+++ src/core/CFConst.gd
@@ -45,6 +45,10 @@ enum ShuffleStyle {
 	OVERHAND,
 }
 
+# When this is set, the player cannot drop cards on the board.
+# If the card is not hovering over a pile when it is dropped, it will
+# simply return to its originating container
+const DISABLE_BOARD_DROP := false
 # Switch this off to disable fancy movement of cards during draw/discard
 const FANCY_MOVEMENT := true
 # The focus style selected for this game. See enum `FocusStyle`
@@ -63,7 +67,7 @@ const PATH_CORE := "res://src/core/"
 # The path where scenes and scripts customized for this specific game exist
 # (e.g. board, card back etc)
 const PATH_CUSTOM := "res://src/custom/"
-# The path where card template scenes exist. 
+# The path where card template scenes exist.
 # These is usually one scene per type of card in the game
 const PATH_CARDS := PATH_CUSTOM + "cards/"
 # The path where the set definitions exist.